Move interal code into internal directories
