+++ src/internal/types.ts
@@ -1,0 +1,4 @@
+export type ReRender = (
+  beforeRender: () => void,
+  afterRender: () => void
+) => void;
